3D Model
- Removed test model
- Added anakin saber model + implement
- Added luke saber model files
