--- src/models/testmodel/license.txt
@@ -1,11 +1,0 @@
-Model Information:
-* title:	[Free] Farm
-* source:	https://sketchfab.com/3d-models/free-farm-7162850a4f0d4bc3a49f3a03f8f87cfe
-* author:	Sampai desu (https://sketchfab.com/Sampai_desu)
-
-Model License:
-* license type:	CC-BY-4.0 (http://creativecommons.org/licenses/by/4.0/)
-* requirements:	Author must be credited. Commercial use is allowed.
-
-If you use this 3D model in your project be sure to copy paste this credit wherever you share it:
-This work is based on "[Free] Farm" (https://sketchfab.com/3d-models/free-farm-7162850a4f0d4bc3a49f3a03f8f87cfe) by Sampai desu (https://sketchfab.com/Sampai_desu) licensed under CC-BY-4.0 (http://creativecommons.org/licenses/by/4.0/)
